+++ crates/api/api/src/federation/list_person_content.rs
@@ -1,4 +1,4 @@
-use crate::stuff::resolve_person_id_from_id_or_username;
+use crate::federation::resolve_person_id_from_id_or_username;
 use activitypub_federation::config::Data;
 use actix_web::web::{Json, Query};
 use lemmy_api_utils::{context::LemmyContext, utils::check_private_instance};
